requirements.txt
discord.py
openai
requests
python-dotenv